refactor: apply big refactor to notification feature
Addapted the big refactor of Firebase on the notification feature.
