--- AFTER-REBASE
@@ -1,1 +1,0 @@
-val notifications: LiveData<List<Notification>> = _notifications
