+++ gradle/libs.versions.toml
@@ -1,5 +1,5 @@
 [versions]
-agp = "8.3.2"
+agp = "8.7.2"
 concurrentFutures = "1.2.0"
 coreTesting = "2.2.0"
 guava = "33.0.0-android"
